Shell out to zuul-cloner instead
We had Zuul installed from source using python 3 since Quibble uses
python3.  That was merely to use it as a library

Since switching to wikimedia-jessie image, we can install Zuul has a
debian package.  Thus, just shell out to zuul-cloner.

The clonemap is dumped as json to a file, yaml being a superset it works
just fine.

Drop requirements.txt since we now solely depends on python built-in
modules.

Change-Id: Ifc158a3025069203f955d029c3cc3e0802d6d438