Add webhost and qunit support
To run the QUnit tests one needs:

$wgEnableJavaScriptTest = true;
npm install
MediaWiki exposed on a webhost
MW_SERVER / MW_SCRIPT to be set to point to the host

Add support for MediaWiki / PHP built-in web server. That is "just"
about invoking php -S with maintenance/dev/includes/router.php
The class DevWebServer is a context manager for sake of simplicity when
it is being used eg later on we would:

 with DevWebServer():
     run_qunit()
     run_selenium()

Php built-in server emits access log and errors to STDERR, send that to
a PIPE read by a Thread() which conveniently relay it to logging.info.

The way karma work, we need wgScriptPath to be empty. Done by passing to
the install --scriptpath= and setting MW_SCRIPT_PATH=''.

FORCE_COLOR='1' forces mocha to use color, quite useful for the Jenkins
console.

Change-Id: I2e67f167d86958680b26d82ab0b17ae3167f9fe7