Disable Chromium sandbox when running in a container
Chromium attempts to set a Linux namespace for PID and network, however
that is not an available capability in a Docker container (at least by
default).  The browser thus refuses to start.

Whenever /.dockerenv exists there are good chance we are in a container.
In such a case pass '--no-sandbox' to CHROMIUM_FLAGS to disable
sandboxing.

Change-Id: I38bc38a74b733b76b9daa288351af95557aef310