Align ChromeDriver and QUnit Chrome arguments
For Chrome we want to pass extra arguments based on the context:

On Docker pass '--no-sandbox'
When no DISPLAY is available pass '--headless'

We use Chrome either via:

* Karma which invokes Chrome directly
* Webdriver IO which relies on Chromedriver responsible for launching
the browser.

Thus we need to factor out setting CHROMIUM_FLAGS so it can be used both
for karma and chromedriver.

Add a few helpers:

quibble.use_headless() based on DISPLAY env variable

quibble.chromium_flags() append appropriate flags to CHROMIUM_FLAGS env
variable based on whether it runs in Docker or has to be headless.

Add tests.

Change-Id: Id292befcda7e4d7935c7a58dae9e0f31084f3b59