tcp_wait: use a different socket on each retry
On python3.6 I noticed the socket get a connection refused and raises an
error, but eventually it ends up connecting in the background. When
retrying the connection that raises an issues about the address already
being in use. The integration test ends up in a deadlock.

Use a new socket on each connection, make sure it is always closed.

Change-Id: Ib5dd648121584593cce533fa9b3c818e2c1f525a