Fix exception when not running 'phpunit'
When not running phpunit tests, Quibble raises:
 Unrecognized zuul_project: mediawiki/core

That is the default value and the variable is used to switch between
PHPUnit testsuite (core, 'extensions', 'skins').  It bails out if it can
not determine the PHPUnit command to run.

When not running phpunit, the logic can be skipped entirely. Hence wrap
the whole block in a single should_run('phpunit').

Reproduction case:

  quibble --run qunit selenium

Change-Id: I3d66ffb04bde48b90113819c5845ff7edc895c06