Also clone $ZUUL_PROJECT repository
Since the Jenkins job already has the ZUUL_PROJECT set, there is no need
to pass it as an argument.  Make set_repos_to_clone() to recognize the
environment variable and add it to the list of repos to clone.

Make sure tests have a clean environment, else when running under CI
ZUUL_PROJECT is set and the actual output contains integration/quibble.

Reorder some bits and prevent core/Vector to be listed twice. Although
zuul-cloner deduplicate the entries, it is less confusing overall.

Change-Id: I68a6e51b4610bb172184f7ba23c281c667ca1002