Allow more recent GitPython version
On a local machine, when using git 2.15, GitPython before 2.1.8 will
fail to recognize an instruction and raise:

  Type of packed-Refs not understood: '# pack-refs with: peeled
  fully-peeled sorted'

That is fixed since GitPython 2.1.8:
https://github.com/gitpython-developers/GitPython/pull/689

We do not want versions 2.1.2 to 2.1.7 that have a performance
regression.

In Docker containers I want to rely on the python3-git python (2.1.1)
over the version from PyPI.  Both Jessie and Stretch ship 2.1.1 and have
older git versions, they are not affected by the issue.

Hence extend the requirement so that it accepts the Debian package
version (2.1.1 < 2.2.0).
Blacklist the versions we do not want (2.1.2 to 2.1.7).
Allow a version from Pypi (2.1.8+) when one uses a recent git.

Bug: T193057
Change-Id: I47367016c496e73fa18ad6d17dc08a39d5d4283d