Support postgres backend
Using pg_virtualenv makes it trivial to start up and connect to a
postgres database. Except that it expects to run a command and then
quit, and doing that with quibble would be backwards.

Instead, quibble starts up pg_virtualenv and has it call a quibble
subcommand that writes the environment variables to a file, and then
waits.

quibble will start running tests, and then teardown the pg_virtulanev
once its done by signaling to the process.

Change-Id: I04dfad8fff681f03a8ff5fec9f253e2677385438