hhvm server should set .svg Content-Type
The MultimediaViewer QUnit tests ends up failling due to a browser error
when loading resources/mmv/img/expand.svg.  The reason is the HHVM
built-in server sends Content-Type: application/octet-stream which
causes the browser to fail loading the image.

HHVM has a map of file/mime type (hhvm.static_file.extensions) which
does not contain svg. Pass the mapping to the command line. Add an
integration tests for both PHP and HHVM.

Bug: T195634
Change-Id: I0cde0812402b400ac18848e189c316757125862e