Prepend mediawiki configuration
We need our custom configuration settings to be available for extensions
when they do not have an extension.json.  And it is probably a good
thing to have the config set before invoking wfLoadExtensions().

The primary use case is Wikibase, its entry point does specific
configuration when running on Jenkins based on the availability of
$wgWikimediaJenkinsCI.  Since our mediawiki configuration snippets were
being appended, the require_once Wikibase.php happened before the
variable is set.

Write our snippets at the top of LocalSettings.php
Adjust _join.php which started the content with '?>' which was meant for
appending the configs.
Update mediawiki.d/README

Bug: T196013
Change-Id: I2d686e78caa385412ce9adc56b5f82f5e82ae48c