Run any command inside Quibble test env
To generate per patch code coverage, all built-in test commands should
be skipped and replaced by a script that generates the coverage. Such a
script would be:

  cd extensions/Foobar
  phpunit-patch-coverage check \
    --command "php -d zend_extension=xdebug.so ../../tests/phpunit/phpunit.php" \
    --html /tmp/coverage-Foobar.html

I still need Quibble to do the installation and spawns the backends.

Add a '--commands' switch that lets one execute commands INSTEAD OF the
built-in stages. Each command runs relatively to MediaWiki installation path.

Bug: T195084
Bug: T195918
Change-Id: I4c85cc2ee43a749147670b5800c7050c05188476