Simply phpunit and allow testsuite override
Introduce --phpunit-testsuite to be able to override a testsuite. The
use case is a mediawiki/core patch that needs to run extensions tests
(wmf-quibble jobs).  That can now be done with:

  --phpunit-testsuite=extensions

Simplify the way we generate the PHPUnit command. The first approach was
based on the type of repos with lot of spaghetti conditionals.  Now, the
testsuite is picked from the command line if available, else for
extensions/skins pick their suite, finally fallback to not selecting any
suite.

For extensions and skins, move the Database group tests at the end of
the build, like we did for mediawiki/core.

Bug: T200017
Change-Id: I77d68f43bbee37a31d2bc714e3350b61ee21ac49