Add support for Zuul --branch and --project-branch
When there is no patchset we default to checking out the patchset branch
via ZUUL_BRANCH.
zuul-cloner has support to specify a different via --branch and
--project-branch (same but on a per project basis).

That will notably be used for DonationInterface. Its master branch has
to be tested with a specific MediaWiki release branch, and vendor to be
pointed to fundraising release branch.

Change-Id: Ie219f2b0871b213a84b8ff7b013e0b76d1c4fb35