In Docker default log dir to be under workspace
The CI jobs do not set --log-dir and Quibble default. We have to mount
both log and src via:

 -v "$(pwd)"/log:/log
 -v "$(pwd)"/src:/workspace

I would like to eventually just mount the Jenkins build workspace
directly as /workspace instead of having to manage several directories
and volumes.

So on the host we would have:

  $WORKSPACE
  ├── log/
  └── src/

And we would run the Docker container with:

 docker run -v "$(pwd):/workspace"

That makes it easy to add other paths as needed.

Requires Wikimedia CI to be adjusted to use --log-dir=/log

Depends-On: I03a7d0a7231d938e71687103f462cd18c9b0499b
Change-Id: If21a336214c30f31caeda8ab73469b16028b2406