mediawiki.d: Improve docs about dev settings and combine env sections
Also remove some comments in the back-compat code path, because this is an
exact copy of the core file. Code comments tend to help maintenance,
but in this case I think was wrongly encouraging people to edit it,
making people forget that it's not actually used in most cases,
because it is only for older MW branches. New changes should be made
to core instead.

Also document since which version, so that we can easily remove it
at some point.

Change-Id: I1d0c099cd96323c91fff8579d117b9caa8c67fd4