Provide option to skip installation of MediaWiki
This is useful if you want to use quibble to checkout the code for a patchset
from Zuul but don't want it to do anything else, example usage would be
`--skip-deps --skip-install`.

The particular use case for this patch is the codehealth pipeline for MediaWiki
core: since at the moment no code coverage is generated either for JS or PHP,
we can save a few minutes by not installing MediaWiki or composer/npm
dependencies. Running `--skip-deps` on its own is not enough since quibble
will still attempt to install MediaWiki, and that will fail since dependencies
are not present.

Change-Id: I875743f2019b2c5b5befd2d2ae9886f0ec4ee47d