Ensure zuul.clone works without mediawiki/core
In a future change I need to clone repos with ZuulCloneCommand() without
listing 'mediawiki/core' as a project to clone. However when running the
cloner in parallel (workers>1), there is an extra check to clone
mediawiki/core first. It fails when it is not listed in the projects to
clone.

Add a very basic test that just attempt to clone repositories in
parallel to ensure no exception get raised.

Change-Id: Iae35e512cb00bf15bbb94cd88bf491ecb8ebea59