Complete test for quibble.zuul.clone()
The method would fail whenever mediawiki/core was in the projects to
list (fixed by 46a2b455e95).

On code review I wanted to test the prepareRepo being made but struggled
with unittest.mock. After reading the mock documentation and testing it
in a sandbox:

Generate the expected calls for the expected repositories using
mock.call() and wrap those with the ThreadPoolExecutor context manager
methods __enter__ and __exit__. This way we have a guarantee we only
prepareRepo our list of repositories.

The list of expected calls:

[
 call(max_workers=2),
 call().__enter__(),
 call().__enter__().submit(
   <MagicMock name='Cloner().prepareRepo'>, 'mediawiki/skins/Foo', ...),
 call().__enter__().submit(
   <MagicMock name='Cloner().prepareRepo', 'mediawiki/skins/Bar', ...),
 call().__exit__(None, None, None)
]

assert_has_calls() then ensure our expected calls is a subset of the
actual calls. And magic happens!

Change-Id: I5be09a190d004e4e70040b3d945785727aece1a0