Exit on cloning failure
When a repository could not be cloned, for example when it does not
exist, the error would be ignored when using parallel cloning. The
reason is the ThreadPoolExecutor futures were not being consumed and
would thus never have their exception bubbled up.

Collect all the futures in a list and invoke result() on each of them as
they complete. That would raise the exception if there is any.

Move the few lines to a standalone function "clone_worker".

Add a thread synchronization between worker to prevent any other
executions as soon as one of the clone fails.  Note that in progress
clones would keep continuing.

Adjust test to mock as_completed()

Bug: T233143
Change-Id: I883b17f178ef4913c2240cde28f0392b1ae2b4b6