Set $wgServer when installing
CLI can not automatically detect the port the server listens on and thus
need to be hinted by setting $wgServer. MediaWiki is also going to drop
support for run-time detection of $wgServer (T232931).

install.php is a maintenance script and as such supports '--server'
which is then recognized to set $wgServer in the generated
LocalSettings.php (confirmed up to REL1_31).

Bug: T233140
Change-Id: I5a3388cf8092d2a9baebfada813340cb99fc499d