qa: use lazy formatting for logging calls
Instead of formatting the message and logging in, delegate that task to
the logging library. Saves some time when the message would not even be
processed, and this means that the context is kept when e.g. logging to
logstash.

Add flake8-logging-format to enforce this, but ignore files in ./zuul/,
since they come from an upstream.

Change-Id: I717e013ed53d988d68397c112442fdf243be003a