Consistent http host
Stop hardcoding 127.0.0.1 has the host to listen on, make it a parameter
to lot of function.

Normalize using a dummy function server_url(host, port) which just
return the http url. Use it to set $wgServer and MW_SERVER which thus
help keeping them in sync.

Add a test to verify whether we support listening on another IP address.
I have picked 127.0.0.2 which is available on the loopback interface.

Change-Id: Ie586c4c9a486edbcc5928679d66ba6b9c377bd3d