Disable color codes around log level words in CI
When executed from Jenkins, disable the color codes for the
line prefixes like "DEBUG:" and "INFO:".

Currently, when the "Console Section" plugin for Jenkins matches
a line, it makes the section start before the text of the current
line but within the color code at the start of the line, which
causes the whole output to be green or blue depending on the
color of the first line that is matched.

Bug: T236222
Change-Id: I50039b229a618f2d094711e9d3c817de27362056