Parallelism as a command object
A composite command which is responsible for running subcommands in
parallel.  These children are spawned as processes.

As the jobs run, a progress report is emitted every 10 seconds
(hardcoded), which looks like:

  Waiting for npm and composer: 40s elapsed, 1/2 completed

This progress monitor is safe to use recursively.

Note that concurrent logging is difficult, and the solution chosen
here comes with trade-offs.  We capture the entire output of each
child process, preventing real-time display so nothing will appear
in the Jenkins console until the producing child thread has finished.
Once a subthread exits, its output is dumped in an atomic `log.info`
call, preventing ambiguous interleaving.  The price we pay is the
loss of real-time monitoring of jobs in the CI console web view.

Doesn't include a usage.  Could be squashed with I6782b184e5d3dd.

Bug: T235449
Bug: T249775
Change-Id: Ib2dc728980ce95dad780f56253596ed03c75d6e9