Split extension and skin npm and composer tests
Npm and composer are independent jobs and have already been running
in parallel, they are implemented as methods in a shared
NpmAndComposer class, and run in parallel using `multiprocessing`
primitives.  This patch reorganizes the code, spliting into separate,
simple commands and wrapping in the new Parallel command.

This is the first use of a Parallel command.

Change-Id: I6782b184e5d3dd3b0b6df76ecd93c1e07de33400