Clone only the target project at first
This saves a huge amount of work which might become unnecessary if
the initial, isolated lint jobs fail.

Since git clone mediawiki/core will fail if the directory is
non-empty, this patch makes additional changes to accomodate:
* If `--skip-zuul` was passed, the only change is that "git clean" in
the project repo is slightly stronger and will also remove any
directories which are their own git repo.  I don't think this can
make a difference in real workflows, since there would be no
opportunity for repos to appear under the project directory.
* If Quibble is responsible for Zuul cloning, then the initial clone
job only checks out the target project, with none of its submodules
or dependencies.
* After single-repository repo linting is finished, Quibble will
delete the project directory (unless this is the core project).  This
is unfortunate but necessary before cloning core.  A fancier
alternative would be to move the project directory aside and then
move it back into place instead of the redundant zuul clone.  Even
without this enhancement, this patch should still be much more
resource-efficient than the status quo.

Bug: T211702
Change-Id: I5ff04ad5b1cdd568146a73140b6515dd4f15ef9a