Docker: Migrate local docker to buster/php73/node10
This means that `docker run -it --rm quibble` from the README now works
again, which is nice (MW master requires PHP 7.2.22, but previous image
was PHP 7.0.33), and massively simplify the build rules for node/npm to
just use upstream, and composer fetching from https://getcomposer.org/,
like we do in the CI build images now.

Other changes:

Updated the list of PHP extensions to match what we actually use in the
CI build images, adding the bcmath, gmp, ldap, and pgsql extensions, as
well as MediaWiki's special wikidiff2 one, and ast for the future so we
can run phan here.

Now setuptools-scm requires that it's run in something that looks to it
like a git repo, so the .dockerignore entry for the .git folder is gone
so that it copies across (and is then dropped). Also, ignore the folder
'src' for local workspace management as recommended exists in the docs.

Added the native unzip and zip utilities, so composer will cease noting
that theoretically php-zip might screw up and break things it downloads
from packagist.

Added ruby and the jsduck gem so that we can run that part of the basic
quibble run without error, too.

Finally, we're now using chromium-driver (renamed from chrome-driver).

Change-Id: Iddaf5f8df3e7b018797538a8e253ed0ccaf20794