Revert "Wipe repo with non-git commands"
This reverts commit 7bfef6e5bdae1f3ba1f82bbc592f4eb7c8059fda.o

That replaces a git clean (which keeps tracked file) with a blind
removal of all files which are really different.

When the workspace already has repositories cloned and using
--skip-zuul the logic tried to wipe files for the triggering project and
eventually fails:

  $ ZUUL_PROJECT=mediawiki/extensions/FileImporter \
      quibble --git-cache=/home/hashar/projects --skip=all
  <...>
  INFO:quibble.commands:>>> Start: Remove directory workspace/src/extensions/FileImporter and any empty parents until workspace/src
  rmdir: failed to remove 'workspace/src/extensions': Directory not empty

When reusing clones, we can't wipe them with `rm -rf` but just need to
get them back to pristine state with `git clean -xqdf`.

Bug: T211702
Change-Id: I2249cf9f5c389c7eb4c5d199d3a617f76a6689a1