tests: no more need to ignore test.py
The testsuite ignore files matching 'test.py' which was to prevent
discovery and execution of non test methods such as "run_composer_test"

The test commands are now in classes and their "*_test" methods are not
discovered by the test suite. Hence we no more need to ignore test.py
which were added with e835d936e8e and 01663f5d9bc.

That is one less issue to have to deal with when migrating from nose to
pytest.

Bug: T254610
Change-Id: I60b353640d2e7decd1758fd8c2e7928c883d70e9