High level testing for the execution plan
I have unfortunately reverted a couple changes by Adam Wight when trying
to optimize the build plan for T211702.

888218fb477 Clone only the target project at first
7bfef6e5bda Wipe repo with non-git commands

The plan did not work when reusing a workspace (--skip-zuul). The
intended behavior can be expressed as a test in tests/test_cmd.py,
specially when we just want to assert a specific command is there with
proper arguments.

Eventually I though it would be good to have some high level acceptance
tests for common usages. They could later on be included in the
documentation.

This change introduces short yaml files to feed environment variables
and command line arguments to Quibble. The test generate the build plan
and assert it is strictly equal to the provided plan.

Bug: T211702
Change-Id: I17bac505580f6a404920cf6e6ac228cbc6a4d1ec