Fix database dumping
Some descendants of BackendServer are not used as context managers and
rely on python garbage collection to do last minute operations (via
__del__).  It might fail when python already deleted other objects or
functions by the team the __del__ is invoked, an example is MySQL which
dies with:

    Traceback (most recent call last):
      File "quibble/quibble/backend.py", in __del__
      File "quibble/quibble/backend.py", in stop
      File "quibble/quibble/backend.py", in dump
    NameError: name 'open' is not defined

The reason is `open` already no more exists which breaks
--dump-db-postrun.

Use a weakref.finalize() to register the shutdown function instead of
__del__. See https://docs.python.org/3/library/weakref.html

Use the same logic on MySQL, Postgresq and DevWebServer.

Note: when used as a context manager, stop() will be called by __exit__
and will be invoked again by weakref.finalize(). But the double
invocation of the code is garded by self.server being set to None and
shortcircuiting the block on a second invocation.

Bug: T239396
Change-Id: I5eafc085b72e61c1e9f1737ec579fc20a3c83f98