Entry point for a CI fullrun
We need a CI job that installs Quibble and does a full run.
utils/ci-fullrun.sh provides the entry point.

It is expected to be run from the root of the repository checkout. It
will then install Quibble and run the tests for mediawiki/core@master.

Note: CI runs this in a Quibble container which already has Quibble
installed. I thus had to instruct python to ignore the site packages by
using `-s` and installing the source to a specific site-packages under
/tmp.

Bug: T235118
Change-Id: I58f72d07e5be0bff34dc29b03da203c01cb3657c