Manage database backend outside of commands
Moves responsibility for the db backend out of the InstallMediaWiki
step and into cmd.py.  Service start is still postponed until just
before it's needed, at the beginning of the integration phase.

The weakref mechanism can be removed, shutdown order is clear now
that the variable is held for the lifetime of top-level QuibbleCmd
and the service is stopped by a context manager.

Bug: T225218
Change-Id: Iee9fb856d7d5f8cca9d45bc7514012466c8dee7b