Delay DatabaseServer initialization
Creating a DatabaseServer instance caused a directory to be created and
the database to be initialized (MySQL, Postgres).  That is now done in
start().

As a result, we can now pass the instance when creating an
InstallMediaWiki and simplify its argument (db_engine, db_dir, dump_dir
are replaced with the database instance).

Removed an obsolete statement about having to hold a reference to the
database instance since it is now hold inside InstallMediaWiki which is
itself hold in the plan list.  We also have a weakref.finalize()
nowadays to ensure it is collected properly.  Antoine has tested the
database postrun step still works properly with:

  quibble --skip-zuul --skip-deps --skip=all --dump-db-postrun

Signed-off-by: Antoine Musso <hashar@free.fr>
Change-Id: I2aa3e137c5fe10204c5f67caa7248f565c004d33