Only run linters for extensions/skins
In bd7c3c94, released with Quibble 0.0.45, we slightly changed the logic
to pick up composer/npm test commands:

    - if quibble.util.isExtOrSkin(zuul_project):
    + if not is_core:
	run_composer = 'composer-test' in stages

For mediawiki/services/parsoid we reuse the same jobs we are using for
MediaWiki extensions. Given it is neither an extension or a skin we were
not running composer test / npm test. Now that is not mediawiki/core we
ran them.

In CI, Parsoid already has dedicated jobs for composer test / npm test,
there is little use to have them run via Quibble as well.

Replace "not is_core" by "(is_extension or is_skin)" which rules out
parsoid.

Add a plan test based on what was produced by 0.0.44.

The extra steps being run for Parsoid were:
+ Extension and skin tests: composer, npm
+ PHPUnit default standalone suite on services/parsoid

Bug: T263500
Change-Id: I60ec80a119a359b5cae7425fb9b701f7a75dff41