Dockerfile launches external apache
The motivation for this patch is to improve performance in processing HTTP
requests during Selenium test runs.  We want to experiment with
apache to see how it compares to the PHP built-in server.

This patch provisions apache and php-fpm using Docker, and the
entrypoint includes flags pointing quibble to the new server.

Bug: T225218
Change-Id: Ife961741a871fa5d29a592aefc8db2f08d5fcefd