Fix Xvfb arguments
The Xvfb arguments were missing a comma, so were silently concatenated
into one parameter. Which means that `-ac` was never taken in account,
that is passed to Xserver which documents it as:

  -ac disables host-based access control mechanisms. Enables access by
      any host, and permits any host to modify the access control list.
      Use with extreme caution. This option exists primarily for running
      test suites remotely.

Since that never got applied, drop it.

The screen 0 and 1280x1024x24 was never in effect either but that is the
default for Xvfb. Anyway explicitly pass them in case we later want to
shrink/enlarge the frame buffer.

Signed-off-by: Antoine Musso <hashar@free.fr>
Change-Id: Ibaf769fc9869b6da1533c0b6ad388ddc3104ae1e