Enhance options passing for pg_virtualenv
We pass all pg_virtualenv options as a single argument which happen to
work but is a bit cumbersome. Instead options passed to pg_createcluster
through `-c` should be an item of the list.

While at it use a long form option for pg_createcluster -s: --socketdir.
Which would also disambiguate with pg_virtualenv -s which is to drop to
a shell on command failure.

Add a test to ensure PostgreSQL can start, marked with 'integration'
since that requires postgresql to be installed.

Change-Id: I4dbfc5d099c70fb3bad15d6fb1253142dc2b20bc