Add support for connecting to already running MySQL
Running `quibble --db-is-external` will attempt to connect to a MySQL
instance available on localhost with the default credentials (username:
root, and no password) used by Quibble. This is useful if you want to
run Quibble with a MySQL db but don't want it to manage
starting/stopping the MySQL backend.

A follow-up commit could provide suport for different dbserver, dbname,
user/password parameters.

Change-Id: I50ca51b0b1fecb3f458207ba0fe7c58206634883