composer: Simplify construction of composer.local.json
If there is a use case for needing only dependencies from
EXT_DEPENDENCIES installed and not the dependencies from all
extensions/skins in the workspace install directory, then this patch
should not be merged. Otherwise, this removes some code and simplifies
running quibble in an existing installation directory, so
EXT_DEPENDENCIES or list of skins/extensions doesn't have to be
constructed as arguments to quibble.

Change-Id: Iab1d465072b0f08dddf6e087b587ed2d0b8b5569