test: fix flappy test for core being cloned first
The test asserts that mediawiki/core is cloned first and check an
extension and a skin get cloned. Due to workers=2, the clone of
extensions and skins are made in parallel and the exact order of the
calls to prepareRepo() might vary.

Assert that the first call is for mediawiki/core which is the primary
intent of this test.

Check that the extension and skin got prepared, but allow any ordering
which does not matter for those.

The issue appeared on CI when running with python3.8.

Change-Id: I2bcd357a55bf11f78055fef8569e1b48e9b0d723