doc: hide the table of content
The main page has a table of content with some maxdepth parameter to
avoid cluttering it. It shows as:

- Quibble CLI usage
    - Positional Arguments
    - Named Arguments
    - stages
- Quibble changelog
- Release check list

The theme already provides a hierarchical navigation in the sidebar
which makes the table of content a bit redundant.

The documents still have to be manually added to the hidden toctree in
`index.rst`. I have considered using a `:glob:` but we would lose the
ability to manual order documents in the navigation bar.

Change-Id: Ieb949defb0791a37b8c188716bec572b009a59b4