doc: move LICENSE to its own document page
Instead of cluttering the main page with licensing term, move that to a
dedicated license document. A link to the license document is shown in
the documentation sidebar.

Change LICENSE to reStructuredText and include it to the license
document. The reStructuredText conversion is borrowed from Jedidiah
Smith at: https://gist.github.com/jedsmith/7852500

Improve some markup.

Change-Id: I620f9be812b0c2da06a389e708e0c9f2fcce4bd0