BrowserTests: Option to run npm install in parallel
Add a command flag to allow for parallelize execution of npm
install for all browser tests at the start of the BrowserTests phase.

The default is to not parallelize, so that this change could be switched
off if it causes problems in CI by updating a job, instead of requiring
a new Quibble version, Docker image build, etc.

Ideally, we'd parallelize the Selenium test execution too, but since
that is going to require quite a bit more work, let's start with the
part that can be done in parallel now.

Note that his has the potential to increase the build time for projects
where a test failure occurs early on (i.e. in core, or in AbuseFilter),
because the tests don't run until npm install completes for all projects
that have browser tests. But for the common scenario where the tests
pass for all repos, this should result in reducing build time.

Bug: T226869
Change-Id: Ifd15768f0f1c9752d4ff530b22b9147c1692cc4e