backend: Check if key exists before deleting
A backend is serialized when it is included in a command that is
parallelized. For those commands, we need to remove the "server" value
from the state, because that contains objects that cannot be
serialized (this causes an error for MySQL for example). But a backend
like SQLite doesn't have a "server" property. So, check if that exists
before attempting to delete it.

Bug: T302226
Change-Id: I9f94136d8acbb9f1a68de17ccc514a4c7f9bc202