ci-fullrun: do not use ZUUL_REF
`$ZUUL_REF` is set by CI and we should not override it.  When set to
`master` it causes zuul.Cloner to find the `master` branch on the
zuul-merger which is not kept in sync with Gerrit. That causes builds to
test the extensions with obsolete code.

Use `quibble --branch=master` instead to indicate Quibble to always
clone the `master` branch of the MediaWiki repository regardless of the
branch name used by `integration/quibble`.

If one want to run against a pending patch of MediaWiki core or an
extensions, they can use `Depends-On` in Quibble and the script will
happilly fetch it via the CI created `$ZUUL_REF`.

Bug: T302707
Change-Id: I5d1487db04f47d0e46918204dd6d1e51cc78b3ab