Define MW_QUIBBLE_CI to replace $wgWikimediaJenkinsCI
Using a constant to indicate that we are running the a specific text
environment follows the pattern established by MW_PHPUNIT_TEST.

A constant is preferrable to a config variable here, since access to
configuration variables that are defined ad-hoc (not defined by MW core)
is unreliable during bootstrap. Depending on how bootstrappign is being
done, they may or may not be available in global scope, may be a global
in the local scope, or only local. Proper access via a Config object is
only possible after bootstrapping is complete.

Change-Id: I0bcc2ffd7638824191b0e63442d89edc86aa9de0
Context: I0f20024803336064ecb07b51fa56581af7b67a85