Handle invalid Unicode from commands
The ProgressReporter reads the commands outputs using a text temporary
file which attempt to decode to unicode the IO stream. We had a case of
a PHPUnit test outputting invalid unicode when asserting the content of
some file headers such as tar/zip.  PHPUnit seems to handle it by
replacing invalid characters with U+FFFD: �.

Change the stream and log collector to use a binary TemporaryFile (the
default). When reading from it set the invalid unicode handler to
'backslashreplace' a backslash followed by the hexadecimal value.

The logging redirect receives a byte stream from the TemporaryFile which
ends up breaking when the handler attempts to `emit()` a string message.
Add a new handler which does the conversion from string to bytes. With
Python 3.8+ this can be replaced by: TemporaryFile(errors="backslashes").

Add test ensuring we cover both stdout and stderr (they are merged
together anyway via `quibble.util.redirect_all_streams`).

The use of multiple context in a with statement confuses black which
ends up crafting a line which is too long. Instruct it to ignore the
faulty line with a `fmt: skip` comment.

Bug: T318029
Change-Id: I9c1c89eb28c6e2bfc1cdcf0a0e9f251cbdfaa01d