Warn if files are left over after git clean -xqdf
For some reason we had a CI build with left over files after running the
composer test stage for an extension. The sequence of commands is
roughly:

- composer test
- git clean -xqdf
- phpunit for the extension

The Phpunit tests discovered tests under a vendor directory under the
extension but they should have been unlinked by git clean. In case this
occurs again, warn with the out of "git status --ignored".

Add a test to cover the logging does occur.

Bug: T321795
Change-Id: Icdef036aedf0be01a48d0837d213e1b785279b89