Allow sending build failure data to external endpoint
Allow Quibble to send data to an HTTP endpoint when a command fails.
Data is sent only if --reporting-url is specified and if the
repository under test has quibble.yaml with a valid config block for
early warnings, e.g.:

```yaml
earlywarning:
  should_vote: 1
  should_comment: 1
```

The environment variable QUIBBLE_API_KEY is sent as a header for
x-api-key; the tool processing data from Quibble can use this to verify
the request.

The use case for this feature is for the earlywarningbot on Toolforge to
comment and/or vote on the relevant patch, alerting developers via
IRC/email/Gerrit as soon as an individual job fails.

Bug: T323750
Change-Id: Ic1303c50403867fec5c578ec0f971aa582d20500