Upgrade GitPython to 3.1.* series
We have long been using GitPython 2.x which was the one used by Zuul and
the Zuul cloner. Some of the releases had issues and we went with a
rather complicated list of requirements to opt out from releases having
issues.

Nowadays, Python 2 is gone and we can move to a newer series.

There are still todo/workaround in the Zuul cloner code which have been
addressed in newer versions GitPython, I am not bothering removing them
since "if it is ain't broke, don't touch it".

Change-Id: I33acab36ecab0ad4d9c0a5a872a3db02ec5bfd52